requirements.txt
fastapi==0.104.1 
uvicorn==0.24.0 
opencv-python-headless==4.8.1.78 
python-multipart==0.0.6 
aiofiles==23.2.1 
python-dotenv==1.0.0 
numpy==1.24.3 
